Add project member invited list endpoint and view
This commit adds a new endpoint and view to handle listing the invited users for a project. The endpoint is `/project/:projectID/invite/list` and the corresponding handler function is `handleProjectMemberInvitedList()`. The view is implemented in the `InvitedUserList` template. This allows users to see the list of invited users for a project and remove invitations if needed.